R: refactor cron with new runtimecontext
