+++ _examples/cron/main.go
@@ -2,19 +2,18 @@ package main
 
 import (
 	"context"
-	"errors"
 	"fmt"
 
-	"github.com/syumai/workers/cloudflare"
 	"github.com/syumai/workers/cloudflare/cron"
 )
 
-func task(ctx context.Context, event *cron.Event) error {
-	fmt.Println(cloudflare.Getenv(ctx, "HELLO"))
-
-	if event.ScheduledTime.Minute()%2 == 0 {
-		return errors.New("even numbers cause errors")
+func task(ctx context.Context) error {
+	e, err := cron.NewEvent(ctx)
+	if err != nil {
+		return err
 	}
+
+	fmt.Println(e.ScheduledTime.Unix())
 
 	return nil
 }
